Rename package `py_vortex` to `vortex`, updating all imports and references
