+++ vortex/pattern.py
